pyproject.toml
[project]
name = "transcriber"
version = "0.1.0"
description = "MCP server for podcast transcription with speaker diarization"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "mlx-whisper>=0.4",
    "pyannote-audio>=3.1",
    "fastmcp>=2.0",
    "torch>=2.0",
    "torchaudio>=2.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1",
    "pytest>=7.0",
]

[project.scripts]
transcriber = "transcriber.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/transcriber"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
